+++ package.json
@@ -51,6 +51,6 @@
     "@types/uuid": "^9.0.8",
     "postcss": "^8",
     "tailwindcss": "^3.4.1",
-    "typescript": "^5"
+    "typescript": "^5.8.3"
   }
 }
